Hyperparameters.jl

Hyperparameters.HYPERPARAMETERS

Hyperparameters.hyperparam

Hyperparameters.hyperparams

Hyperparameters.report_hyperparameters

Hyperparameters.HYPERPARAMETERS — Constant

HYPERPARAMETERS::Dict{Symbol, Any}

Collection of all the hyperparameters, and their values accessed during this run.

Hyperparameters.hyperparam — Method

hyperparam([T::Type=Float64,] name; prefix="SM_HP_"))

Load the hyperparameter with the given name from the enviroment variable named with the name in uppercase, and prefixed with prefix parsing it as type T (default: Float64).

Also stores the hyperparameter and its value in the global HYPERPARAMETERS dictionary. This function is generally expected to be used with SageMaker, and supplies the default prefix for it.

using Hyperparameters
ENV["HP_POWER_LEVEL"] = "9001"
hyperparam(:power_level; prefix="HP_")

# output
9001.0

Hyperparameters.hyperparams — Method

hyperparams(names...; prefix="SM_HP_")

As per hyperparam, but taking multiple names and returning a NamedTuple.

using Hyperparameters
ENV["SM_HP_A"] = "5"
ENV["SM_HP_B"] = "1.22"
hyperparams(:a, :b, types=[Int, Float64])

# output
(a = 5, b = 1.22)

Also stores the hyperparameters and their values in the global HYPERPARAMETERS dictionary.

Hyperparameters.report_hyperparameters — Method

report_hyperparameters(save_dir::AbstractPath)

Saves the cached HYPERPARAMETERS to a JSON file named "hyperparameters.json" in the save_dir and prints each key-value pair to the logger.

The regex to extract the components is: hyperparameters: (?<key>)=(?<value>).